add issue cataloger
adds an env called `tune` to tune's namespace which allows for flexibly interfacing with information about the current run. the environment is re-initialized with each call to `fit_resamples()` or `tune_*()`, so each user-facing run gets a fresh issue cataloger. the environment contains a "catalog," which is similar to `out_notes` but contains the count of issues in _previous steps_ in the grid loop. comparing the `catalog` (from the `tune` env) with the up-to-date `issues` allows us to determine which issues are "new" and need addition to the progress bar.