generate needed objects from `split` first-thing (#910)
Generates the analysis, potato, and assessment sets first-thing in tune_grid_loop_iter(), as well as needed labels and prediction indices, and then no longer needs to reference split for the rest of the function body. As such, refers to those sets only by their names and removes references to split and split_orig in explanatory comments.